rabbit/amqp/requirements.txt
pika==1.2.0
python-dotenv==0.20.0
